Move code in rest-jackson-common to a proper package
Fixes: #44740
