+++ extensions/resteasy-reactive/rest-jackson-common/runtime/src/main/java/io/quarkus/resteasy/reactive/jackson/common/runtime/serialisers/vertx/VertxJsonArrayBasicMessageBodyReader.java
@@ -1,4 +1,4 @@
-package io.quarkus.resteasy.reactive.jackson.runtime.serialisers.vertx;
+package io.quarkus.resteasy.reactive.jackson.common.runtime.serialisers.vertx;
 
 import java.io.IOException;
 import java.io.InputStream;
